requirements.txt
panel
langchain
langchain_experimental
pandas
plotly
python-dotenv
langchain_community